backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
mistralai>=0.4.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3
opencv-python-headless>=4.9.0
weave>=0.50.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pydantic>=2.0.0
elevenlabs